### 2026-01-27 #280 @tradertanmay
TanML
Submitting Author: @tradertanmay
All current maintainers: @tradertanmay
Package Name: tanml
One-Line Description of Package: Automated Model Validation Toolkit for Tabular Machine Learning
Repository Link:  https://github.com/tdlabs-ai/tanml
Version submitted: 0.1.10
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

    Data validation & testing: TanML automates tabular data checks (profiling, drift) to ensure data quality in ML workflows.

    Workflow automation: TanML streamlines end-to-end model validation and generates audit-ready reports for reproducible governance.

    Target audience: Data scientists and ML engineers building tabular models, plus model validation/MRM teams who need standardized,
     repeatable checks.

     Scientific applications: Reproducible evaluation of tabular ML models and datasets (data quality, performance, robustness, fairness,
     explainability) with automated report generation for research and audit documentation.

  - Are there other Python packages that accomplish the same thing? If so, how does yours differ? None

- [x ] contains a README with instructions for installing the development version.
- [x ] contains a tutorial with examples of its essential functions and uses.
- [x ] has continuous integration setup, such as GitHub Actions CircleCI, and/or others.

- [x ] I have read the [author guide](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html).

- [ x] [Last but not least please fill out our pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us track
