### 2026-02-06 c2927ca
docs & ci: address pyOpenSci review comments and reorganize local assets
- Integrated ruff and mypy into CI pipeline
- Added testing instructions and documentation link to README
- Isolated local research, manuscripts, and archives to local_research/
- Cleaned root directory to match public repository structure
