### 2026-02-06 ec4b99f
ci: fix final mypy type validation and reformat codebase
