+++ tanml/ui/services/cv.py
@@ -7,21 +7,32 @@ This module contains the _run_repeated_cv function extracted from app.py.
 
 from __future__ import annotations
 
+
 def _run_repeated_cv(model, X, y, task_type, n_splits=5, n_repeats=5, seed=42):
     """
     Runs RepeatedStratifiedKFold (clf) or RepeatedKFold (reg) and returns
     a dict of stats suitable for report injection (mean, std, p05, p50, p95).
     """
-    from sklearn.model_selection import RepeatedStratifiedKFold, RepeatedKFold
+    import numpy as np
     from sklearn.base import clone
     from sklearn.metrics import (
-        roc_auc_score, f1_score, precision_score, recall_score,
-        accuracy_score, matthews_corrcoef, balanced_accuracy_score,
-        mean_squared_error, mean_absolute_error, r2_score, median_absolute_error,
-        log_loss, brier_score_loss, average_precision_score
+        accuracy_score,
+        average_precision_score,
+        balanced_accuracy_score,
+        brier_score_loss,
+        f1_score,
+        log_loss,
+        matthews_corrcoef,
+        mean_absolute_error,
+        mean_squared_error,
+        median_absolute_error,
+        precision_score,
+        r2_score,
+        recall_score,
+        roc_auc_score,
     )
+    from sklearn.model_selection import RepeatedKFold, RepeatedStratifiedKFold
     from sklearn.preprocessing import LabelEncoder
-    import numpy as np
 
     # Decide CV strategy
     if task_type == "classification":
@@ -29,10 +40,10 @@ def _run_repeated_cv(model, X, y, task_type, n_splits=5, n_repeats=5, seed=42):
         # Ensure y is proper for stratification
         # If string labels, encode them temporarily for CV split generation
         try:
-             le = LabelEncoder()
-             y_enc = le.fit_transform(y)
+            le = LabelEncoder()
+            y_enc = le.fit_transform(y)
         except:
-             y_enc = y
+            y_enc = y
     else:
         cv = RepeatedKFold(n_splits=n_splits, n_repeats=n_repeats, random_state=seed)
         y_enc = y
@@ -42,24 +53,24 @@ def _run_repeated_cv(model, X, y, task_type, n_splits=5, n_repeats=5, seed=42):
 
     def _add(k, v):
         scores.setdefault(k, []).append(v)
-    
+
     # Pre-check if model handles probabilities
     has_proba = hasattr(model, "predict_proba")
 
-    X_arr = X.values if hasattr(X, 'values') else np.array(X)
-    y_arr = y.values if hasattr(y, 'values') else np.array(y)
-    
+    X_arr = X.values if hasattr(X, "values") else np.array(X)
+    y_arr = y.values if hasattr(y, "values") else np.array(y)
+
     # Ensure arrays are numeric (handle edge cases with SAS/SPSS data)
     try:
         X_arr = X_arr.astype(np.float64)
     except (ValueError, TypeError):
         # If conversion fails, try to select only numeric columns
-        if hasattr(X, 'select_dtypes'):
+        if hasattr(X, "select_dtypes"):
             X_numeric = X.select_dtypes(include=[np.number])
             X_arr = X_numeric.values.astype(np.float64)
         else:
             X_arr = np.array(X, dtype=np.float64)
-    
+
     # Handle y conversion
     try:
         y_arr = y_arr.astype(np.float64)
@@ -70,112 +81,155 @@ def _run_repeated_cv(model, X, y, task_type, n_splits=5, n_repeats=5, seed=42):
     for train_idx, test_idx in cv.split(X_arr, y_enc):
         X_tr, X_val = X_arr[train_idx], X_arr[test_idx]
[…]
