### 2026-02-06 5bff323
ci: disable mypy return-value error for python 3.10 compatibility
