### 2026-02-06 #280 @tradertanmay
Hi @crhea93, thank you for the feedback.
#### Automated Unit Testing
- Implemented a GitHub Actions workflow that automatically runs unit tests on every push.
- The tests run across Python versions 3.10 through 3.13 to ensure compatibility.
#### Local Testing and Dependencies
- Added a Testing section to the README with instructions for local verification.
- Configured development dependencies so they can be installed with a single command: pip install -e ".[dev]".
#### Quality Tooling and Type Checks
- Integrated Ruff for automated linting and code formatting in the CI pipeline.
- Implemented Mypy for static type checking across the library.
#### Documentation Discoverability
- Moved the documentation link to the top of the README for immediate visibility.

**Status**: All automated checks for linting, type safety, and unit testing are currently passing in the GitHub repository.
