### 2026-02-22 #280 @anabeltan
Hi @tradertanmay , and thanks @crhea93 for coordinating,

Thanks for submitting your package to pyOpenSci! Please find my comments below.

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

Overall, this is a strong submission. The repository is well structured, the documentation is generally clear, and the test suite runs successfully (24 passed, 1 warning).
```
============================================== warnings summary ===============================================
tests/test_analysis_metrics.py::test_vif_calculation
  /home/usr/review/.venv/lib/python3.10/site-packages/statsmodels/stats/outliers_influence.py:197: RuntimeWarning: divide by zero encountered in scalar divide
    vif = 1. / (1. - r_squared_i)

======================================= 24 passed, 1 warning in 11.70s ========================================
```

Below are a few minor suggestions for improvement:

1. Table of Contents Section

The README would benefit from adding clickable links in the Table of Contents. Currently, the sections are listed but not hyperlinked. Converting them into links would allow readers to quickly nagvigate to the relevant sections, significantly improving usability.

2. UI Access via ```tanml ui```

When running ```tanml ui```, the process outputs:

```
You can now view your Streamlit app in your browser.

```

However, the UI was not accessible via the browser in my setup. This is likely due to running Linux remotely from Windows using SSH. Since ```127.0.0.1``` refers to the remote machine in this context, it is not directly accessible from the local Windows browser without proper port forwarding. It would be very helpful to include a short section in the README explaining how to use the software over SSH. A brief "Running the UI on a remote server" subsection would greatly improve accessibility for users working in cluster or SSH-based environments.

3. Testing Section

In the Testing section of the README, one small but important step is missing: ```cd ./tanml/```.

While this may seem trivial for experienced users, explicitly stating this step would improve clarity for users who are less familiar with Python project layouts or repository structures.
