### 2026-02-26 #280 @tradertanmay
Hi @anabeltan, thank you for your review and for your valuable suggestions. And thank you @crhea93  for serving as the editor.

I truly appreciate the time and effort you put into the feedback. I will wait for Reviewer 2’s comments. Once I receive feedback, I will incorporate all suggestions.

In parallel, I am preparing a paper.md file that aligns with JOSS submission requirements.
