### 2026-03-05 #280 @echen1214
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

    - [ ] Python versions supported

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

Estimated hours spent reviewing: ~8 hours over the course of two sessions

---

Overall the package worked as expected when tested on a small .csv file and is a strong submission.

Major notes:
        - one touted aspect of the TanML model is that it is fully UI-driven but the documentation on the UI is a bit lacking and there aren't any tutorials, or example datasets on the UI
        - some aspects of the UI are challenging to use. for example when i use the app...
                - the main UI clicking on workflow are not functional.
                - in dark mode, the choice of font color is the same as the background color making text illegible. this is corrected when using light mode

Minor notes:
        - when running the tests i get the same results as @anabeltan
        - i had trouble pip installing tanml with python3.10 due failing to build llvmlite. but this was resolved by first installing `pip installing numba==0.59.0`. didn't look too deeply into how to resolve this but it could help future developers.
        - the python version is listed but not as a badge
        - in the contributing portion -- need to add a line to ```git clone https://github.com/tdlabs-ai/tanml.git```
        - the API reference documentation could also fleshed out. For example, ```build_estimator``` is listed in the tutorial but there isn't any corresponding documentation
