### 2026-03-09 248999e
feat: implement Phase 3 core analytical features and address pyOpenSci feedback
