+++ paper.bib
@@ -1,0 +1,110 @@
+@software{tanml_2025,
+  author       = {Sah, Tanmay and Sah, Dolly},
+  title        = {TanML: Automated Model Validation Toolkit for Tabular Machine Learning},
+  year         = {2025},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.17317165},
+  url          = {https://doi.org/10.5281/zenodo.17317165}
+}
+
+@article{scikit-learn,
+  author  = {Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and VanderPlas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'E}douard},
+  title   = {Scikit-learn: Machine Learning in Python},
+  journal = {Journal of Machine Learning Research},
+  year    = {2011},
+  volume  = {12},
+  number  = {85},
+  pages   = {2825--2830},
+  url     = {https://www.jmlr.org/papers/v12/pedregosa11a.html}
+}
+
+@inproceedings{xgboost,
+  author    = {Chen, Tianqi and Guestrin, Carlos},
+  title     = {XGBoost: A Scalable Tree Boosting System},
+  booktitle = {Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  year      = {2016},
+  pages     = {785--794},
+  doi       = {10.1145/2939672.2939785},
+  url       = {https://arxiv.org/abs/1603.02754}
+}
+
+@inproceedings{lightgbm,
+  author    = {Ke, Guolin and Meng, Qi and Finley, Thomas and Wang, Taifeng and Chen, Wei and Ma, Weidong and Ye, Qiwei and Liu, Tie-Yan},
+  title     = {LightGBM: A Highly Efficient Gradient Boosting Decision Tree},
+  booktitle = {Advances in Neural Information Processing Systems 30},
+  year      = {2017},
+  url       = {https://papers.nips.cc/paper/6907-lightgbm-a-highly-efficient-gradient-boosting-decision-tree}
+}
+
+@inproceedings{catboost,
+  author    = {Prokhorenkova, Liudmila and Gusev, Gleb and Vorobev, Aleksandr and Dorogush, Anna Veronika and Gulin, Andrey},
+  title     = {CatBoost: unbiased boosting with categorical features},
+  booktitle = {Advances in Neural Information Processing Systems 31},
+  year      = {2018},
+  url       = {https://proceedings.neurips.cc/paper/2018/hash/14491b756b3a51daac41c24863285549-Abstract.html}
+}
+
+@software{streamlit,
+  author   = {{Streamlit, Inc.}},
+  title    = {Streamlit},
+  year     = {2026},
+  url      = {https://github.com/streamlit/streamlit},
+  note     = {Open-source Python framework for building interactive data applications}
+}
+
+@software{ydata-profiling,
+  author   = {{YData}},
+  title    = {ydata-profiling},
+  year     = {2026},
+  url      = {https://github.com/ydataai/ydata-profiling},
+  note     = {Open-source data profiling and exploratory data analysis library}
+}
+
+@software{evidently,
+  author   = {{Evidently AI}},
+  title    = {Evidently},
+  year     = {2026},
+  url      = {https://github.com/evidentlyai/evidently},
+  note     = {Open-source framework to evaluate, test, and monitor ML and LLM-powered systems}
+}
+
+@software{deepchecks,
+  author   = {{Deepchecks}},
[…]
