+++ paper2.md
@@ -1,0 +1,92 @@
+---
+title: 'TanML: Automated Model Validation Toolkit for Tabular Machine Learning'
+tags:
+  - Python
+  - machine learning
+  - model validation
+  - model risk management
+  - regulatory compliance
+  - governance
+authors:
+  - name: Tanmay Sah
+    orcid: 0009-0004-8583-2208
+    affiliation: 1
+  - name: Dolly Sah
+    affiliation: 2
+  - name: Harshul Jain
+    affiliation: 3
+  - name: Tanya Sah
+    affiliation: 3
+  - name: Kayden Jordan
+    affiliation: 1
+affiliations:
+  - name: Harrisburg University of Science and Technology
+    index: 1
+  - name: University of Utah
+    index: 2
+  - name: New York University
+    index: 3
+date: 7 March 2026
+bibliography: paper.bib
+---
+
+# Summary
+
+TanML [@tanml_2025] is an open-source toolkit for validating tabular machine learning models through a user-friendly Streamlit interface. It supports an end-to-end workflow that includes data profiling, preprocessing, feature ranking, model development, evaluation, and export of editable audit-ready Word reports. The toolkit is designed to help users validate models built with common Python ML libraries scikit-learn [@scikit-learn], XGBoost [@xgboost], LightGBM [@lightgbm], and CatBoost [@catboost]. TanML is especially useful in regulated and documentation-heavy settings because it combines model validation, explainability, drift and stress analysis, and report generation in a single local-first workflow. It is intended for data scientists, quantitative analysts, and model risk stakeholders who need a practical way to review and document tabular ML models without building custom validation pipelines from scratch.
+
+# Statement of need
+
+Tabular machine learning models are widely used in high-stakes decision settings such as finance, insurance, and other regulated domains. In these environments, model development alone is insufficient, because practitioners must also evaluate data quality, feature behavior, predictive performance, robustness, and interpretability while producing documentation to support internal review, governance, and compliance processes. In practice, these activities are often fragmented across notebooks, scripts, visualization tools, and manually prepared reports, making validation workflows difficult to standardize, reproduce, and communicate. TanML was developed to address this problem by providing an integrated validation workflow for tabular machine learning models. The software brings together data profiling, preprocessing, feature ranking, model development, model evaluation, explainability, and export of editable audit-ready reports within a single interface, reducing the operational gap between model building and model validation, especially for users who need structured outputs rather than ad hoc analyses. The primary target audience for TanML includes data scientists, quantitative analysts, model validation teams, and practitioners working in regulated or documentation-heavy settings. The software is particularly useful for users who require reviewable validation artifacts, not only numerical metrics or exploratory dashboards, and its UI-driven workflow lowers the barrier for domain experts and stakeholders who may not wish to implement custom validation pipelines directly in code. TanML relates to several existing categories of software, but its focus differs from each of them: data profiling tools such as ydata-profiling support exploratory analysis of datasets, monitoring frameworks such as Evidently focus on drift and production observability, validation libraries such as Deepchecks provide programmable checks for data and models, and AutoML frameworks such as PyCaret and AutoGluon emphasize model training and selection. TanML’s contribution lies in integrating these validation-oriented needs into a unified workflow centered on tabular ML validation and stakeholder-ready documentation, making it a complementary research software package that fills the gap between model development utilities and governance-oriented validation practice.
+
+
+# State of the field
+
[…]

[…]
