### 2026-03-09 1ce0f5c
fix: update JOSS paper action name
