+++ .github/workflows/paper.yml
@@ -8,7 +8,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
       - name: Build PDF
-        uses: openjournals/openjournals-multi-journal-action@master
+        uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
           paper_path: paper.md
