### 2026-03-09 24fbd4d
fix: use hyphen in paper-path workflow input
