### 2026-03-09 a6a3e2f
fix: consolidate paper workflows and add explicit artifact upload
