+++ .github/workflows/joss-paper.yml
@@ -28,3 +28,9 @@ jobs:
         with:
           journal: joss
           paper-path: paper.md
+
+      - name: Upload PDF
+        uses: actions/upload-artifact@v4
+        with:
+          name: paper
+          path: paper.pdf
