+++ .github/workflows/paper.yml
@@ -1,14 +1,0 @@
-name: Generate PDF
-on: [push, pull_request]
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          paper-path: paper.md
