### 2026-03-09 8b1bcbe
fix: use PNG for architecture diagram to ensure crisp rendering in JOSS PDF
