### 2026-03-09 f86ebf6
fix: remove redundant Figure 1 prefix in caption
