### 2026-03-09 fab98f3
fix: remove image alt-text to prevent double Figure labeling in JOSS PDF
