### 2026-03-09 c3b89d4
fix: restore figure caption with label and update trigger paths
