### 2026-03-10 8c87a82
fix: format table 1 with correct markdown caption syntax
