### 2026-03-10 11f6538
chore: remove unwanted paper.pdf and paper2.md from tracking
