### 2026-03-10 55e0d88
fix: format table 1 caption with blank line and label for pandoc
