### 2026-03-10 dc143fa
fix: move table caption to the bottom of the markdown table
