### 2026-03-10 33e9e37
fix: restore full comparison table and proper pandoc caption syntax
