### 2026-03-10 1916c20
fix: attach pandoc table caption directly to table bottom without blank line
