### 2026-03-10 65ae1ac
fix: convert comparison table to pandoc grid table for proper rendering
