### 2026-03-10 68d5ce8
docs: add upgrade command to installation instructions
