+++ README.md
@@ -75,6 +75,9 @@ source .venv/bin/activate  # Linux/Mac
 
 # Install TanML
 pip install tanml
+
+# Upgrade an existing installation
+python -m pip install --upgrade tanml
 ```
 
 ## Run TanML Locally
