### 2026-03-10 2312198
docs: update changelog with phase 3 and joss publication features
