### 2026-03-10 e1a670f
docs: remove incorrect phase 3 features from changelog
