### 2026-03-10 e61cc7d
docs: correct typo in 0.1.10 release date from 2024 to 2026
