### 2026-03-10 f934e40
docs: add pyOpenSci review fixes to changelog
