### 2026-03-10 d34f6fd
docs: remove incorrect SHAP feature from changelog
