+++ CHANGELOG.md
@@ -30,5 +30,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Reporting**: Initiated refactor of Word report generators to improve maintainability.
 - Added Python 3.13 to the official testing matrix.
 
-## [0.1.10] - 2026-03-07
+## [0.1.10] - 2026-01-25
 - Initial baseline for pyOpenSci submission.
