### 2026-03-10 37177ce
docs: set release date for version 0.2.0
