### 2026-03-10 5d3e5c6
docs: add dark mode legibility fix to changelog
