### 2026-03-10 a90fa30
chore: bump version to 0.2.0
