### 2026-03-10 70cb658
docs: refine JOSS paper bibliography with DOIs and core stack citations
