+++ paper.bib
@@ -1,7 +1,7 @@
-@software{tanml_2025,
+@software{tanml_2026,
   author       = {Sah, Tanmay and Sah, Dolly},
   title        = {TanML: Automated Model Validation Toolkit for Tabular Machine Learning},
-  year         = {2025},
+  year         = {2026},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.17317165},
   url          = {https://doi.org/10.5281/zenodo.17317165}
@@ -29,10 +29,11 @@
 }
 
 @inproceedings{lightgbm,
-  author    = {Ke, Guolin and Meng, Qi and Finley, Thomas and Wang, Taifeng and Chen, Wei and Ma, Weidong and Ye, Qiwei and Liu, Tie-Yan},
+  author    = {Ke, Guolin and Meng, Qi slogans and Finley, Thomas and Wang, Taifeng and Chen, Wei and Ma, Weidong and Ye, Qiwei and Liu, Tie-Yan},
   title     = {LightGBM: A Highly Efficient Gradient Boosting Decision Tree},
   booktitle = {Advances in Neural Information Processing Systems 30},
   year      = {2017},
+  doi       = {10.5555/3294996.3295074},
   url       = {https://papers.nips.cc/paper/6907-lightgbm-a-highly-efficient-gradient-boosting-decision-tree}
 }
 
@@ -41,6 +42,7 @@
   title     = {CatBoost: unbiased boosting with categorical features},
   booktitle = {Advances in Neural Information Processing Systems 31},
   year      = {2018},
+  doi       = {10.5555/3327757.3327770},
   url       = {https://proceedings.neurips.cc/paper/2018/hash/14491b756b3a51daac41c24863285549-Abstract.html}
 }
 
@@ -56,6 +58,7 @@
   author   = {{YData}},
   title    = {ydata-profiling},
   year     = {2026},
+  doi      = {10.5281/zenodo.10123516},
   url      = {https://github.com/ydataai/ydata-profiling},
   note     = {Open-source data profiling and exploratory data analysis library}
 }
@@ -72,6 +75,7 @@
   author   = {{Deepchecks}},
   title    = {Deepchecks},
   year     = {2024},
+  doi      = {10.5281/zenodo.6344777},
   url      = {https://github.com/deepchecks/deepchecks},
   note     = {Open-source framework for continuous validation of machine learning models and data}
 }
@@ -89,6 +93,7 @@
   title   = {AutoGluon-Tabular: Robust and Accurate AutoML for Structured Data},
   journal = {arXiv preprint arXiv:2003.06505},
   year    = {2020},
+  doi     = {10.48550/arXiv.2003.06505},
   url     = {https://arxiv.org/abs/2003.06505}
 }
 
@@ -107,4 +112,38 @@
   month        = apr,
   url          = {https://www.federalreserve.gov/supervisionreg/srletters/sr1107.htm},
   note         = {Issued April 4, 2011}
+}
+
+@article{numpy,
+  title     = {Array programming with {NumPy}},
[…]
