### 2026-03-10 adb8996
docs: sync paper.md citations with updated bib keys
