### 2026-03-10 528b7c7
docs: fix corrupted grid table in paper.md
