### 2026-03-10 61becc0
docs: restore perfectly aligned pandoc grid table
