### 2026-03-10 b254806
docs: reinstate original successful grid table format
