### 2026-03-10 13137bc
docs: fix pandoc grid table parsing with preceding blank line
