### 2026-03-10 d14d632
docs: revert bibtex keys to working state while keeping grid table format
