### 2026-03-10 8b702bb
docs: resolve citation mismatch by syncing paper.md to use the original short keys present in paper.bib
