### 2026-03-10 a9e51ad
docs: use exact Google Scholar citations as intended by the user
