+++ paper.bib
@@ -7,7 +7,7 @@
   url          = {https://doi.org/10.5281/zenodo.17317165}
 }
 
-@article{scikit-learn,
+@article{pedregosa2011scikit,
   author  = {Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and VanderPlas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'E}douard},
   title   = {Scikit-learn: Machine Learning in Python},
   journal = {Journal of Machine Learning Research},
@@ -18,7 +18,7 @@
   url     = {https://www.jmlr.org/papers/v12/pedregosa11a.html}
 }
 
-@inproceedings{xgboost,
+@inproceedings{chen2016xgboost,
   author    = {Chen, Tianqi and Guestrin, Carlos},
   title     = {XGBoost: A Scalable Tree Boosting System},
   booktitle = {Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
@@ -28,7 +28,7 @@
   url       = {https://arxiv.org/abs/1603.02754}
 }
 
-@inproceedings{lightgbm,
+@inproceedings{ke2017lightgbm,
   author    = {Ke, Guolin and Meng, Qi slogans and Finley, Thomas and Wang, Taifeng and Chen, Wei and Ma, Weidong and Ye, Qiwei and Liu, Tie-Yan},
   title     = {LightGBM: A Highly Efficient Gradient Boosting Decision Tree},
   booktitle = {Advances in Neural Information Processing Systems 30},
@@ -37,7 +37,7 @@
   url       = {https://papers.nips.cc/paper/6907-lightgbm-a-highly-efficient-gradient-boosting-decision-tree}
 }
 
-@inproceedings{catboost,
+@inproceedings{prokhorenkova2018catboost,
   author    = {Prokhorenkova, Liudmila and Gusev, Gleb and Vorobev, Aleksandr and Dorogush, Anna Veronika and Gulin, Andrey},
   title     = {CatBoost: unbiased boosting with categorical features},
   booktitle = {Advances in Neural Information Processing Systems 31},
@@ -54,7 +54,7 @@
   note     = {Open-source Python framework for building interactive data applications}
 }
 
-@software{ydata-profiling,
+@software{clemente2023ydata,
   author   = {{YData}},
   title    = {ydata-profiling},
   year     = {2026},
@@ -88,7 +88,7 @@
   note     = {Open-source, low-code machine learning library in Python}
 }
 
-@article{autogluon,
+@article{erickson2020autogluon,
   author  = {Erickson, Nick and Mueller, Jonas and Shirkov, Alexander and Zhang, Hang and Larroy, Pedro and Li, Mu and Smola, Alexander J.},
   title   = {AutoGluon-Tabular: Robust and Accurate AutoML for Structured Data},
   journal = {arXiv preprint arXiv:2003.06505},
@@ -97,7 +97,7 @@
   url     = {https://arxiv.org/abs/2003.06505}
 }
 
-@inproceedings{shap,
+@inproceedings{lundberg2017unified,
   author    = {Lundberg, Scott M. and Lee, Su-In},
   title     = {A Unified Approach to Interpreting Model Predictions},
   booktitle = {Advances in Neural Information Processing Systems 30},
@@ -114,7 +114,7 @@
   note         = {Issued April 4, 2011}
 }
 
-@article{numpy,
[…]
