### 2026-03-10 dfc4a59
docs: refine comparison table formatting and remove citations
