### 2026-03-10 17041be
docs: remove specific AutoML framework names from comparison table
