### 2026-03-10 b682f59
docs: user-authored refinements to paper text and comprehensive acknowledgements
