### 2026-03-10 c8b3e87
docs: finalize reporting terminology in paper and workflow section
