### 2026-03-10 468272f
docs: fix markdown header rendering for References section
