### 2026-03-10 a2acb65
docs: remove beta status from README to reflect project maturity
