### 2026-03-10 f962a4b
docs: sync citation year with JOSS paper (2025 -> 2026)
