### 2026-03-10 9e12eca
style: unify python badge styling in README
