### 2026-03-10 008080d
docs: sync home page with README improvements
