### 2026-03-10 29d02f3
docs: unify aesthetics and content between documentation portal and README
