### 2026-03-10 daf5354
docs: revert theme and aesthetic changes as requested
