### 2026-03-10 39db56f
docs: automate README to docs/index.md synchronization in deployment workflow
