### 2026-03-13 1a532df
docs: final refinements to paper text and bibliography for JOSS submission
