+++ paper.md
@@ -32,16 +32,16 @@ bibliography: paper.bib
 
 # Summary
 
-TanML [@tanml_2026] is an open-source toolkit for validating tabular machine learning models through a user-friendly Streamlit interface. It supports an end-to-end workflow that includes data profiling, preprocessing, feature ranking, model development, evaluation, and export of editable audit-ready reports. The toolkit is designed to help users validate models built with common Python ML libraries such as NumPy [@harris2020array], Pandas [@reback2020pandas], scikit-learn [@pedregosa2011scikit], XGBoost [@chen2016xgboost], LightGBM [@ke2017lightgbm], and CatBoost [@prokhorenkova2018catboost]. TanML is especially useful in regulated and documentation-heavy settings [@sr11-7] because it combines model validation, explainability, drift and stress analysis, and report generation in a single local-first workflow. It is intended for data scientists, quantitative analysts, and model risk stakeholders who need a practical way to review and document tabular ML models without building custom validation pipelines from scratch.
+TanML [@tanml_2026] is an open-source toolkit for validating machine learning models built on tabular (structured) data through a user friendly Streamlit interface. It supports an end-to-end workflow that includes data profiling, preprocessing, feature ranking, model development, evaluation, and export of editable audit ready reports. The toolkit is designed to help users validate models built with common Python ML libraries such as scikit-learn [@pedregosa2011scikit], XGBoost [@chen2016xgboost], statsmodels [@seseabold2010statsmodelsa], LightGBM [@ke2017lightgbm], and CatBoost [@prokhorenkova2018catboost]. TanML is especially useful in regulated and documentation heavy settings [@sr11-7] because it combines model validation, explainability, drift and stress analysis, and report generation in a single local first workflow. It is intended for data scientists, quantitative analysts, and model risk stakeholders who need a practical way to review and document ML models trained on structured datasets without building custom validation pipelines from scratch.
 
 # Statement of need
 
-Tabular machine learning models are widely used in high-stakes domains such as finance, insurance, and other regulated settings. In these environments, building a predictive model is only one part of the workflow, because practitioners must also assess data quality, model behavior, predictive performance, robustness, and interpretability while producing documentation that supports internal review, governance, and compliance processes. In practice, these validation activities are often fragmented across notebooks, scripts, visualization libraries, and manually prepared reports, making workflows difficult to standardize, reproduce, and communicate, especially when results must be reviewed by stakeholders beyond the original model developer. TanML was developed to address this problem by providing a unified workflow for tabular model validation and documentation. Instead of requiring practitioners to assemble separate tools for analysis, evaluation, and reporting, TanML organizes these activities within a single interface intended to make validation outputs easier to reproduce, review, and communicate. The primary target audience includes data scientists, quantitative analysts, model validation teams, and other practitioners working in regulated or documentation-heavy settings who need reviewable validation artifacts in addition to numerical metrics or exploratory analysis. TanML relates to several existing categories of software, including data profiling tools, monitoring frameworks, validation libraries, and AutoML systems, but these tools typically emphasize only one portion of the workflow, such as exploratory analysis, production monitoring, programmable checks, or model training. TanML is designed to address the gap between model development utilities and governance-oriented validation workflows for tabular machine learning.
[…]
