+++ paper.bib
@@ -147,3 +147,13 @@
   doi       = {10.1109/MCSE.2007.55},
   publisher = {IEEE Computer Society}
 }
+
+@article{seabold2010statsmodels,
+  title={Statsmodels: econometric and statistical modeling with python.},
+  author={Seabold, Skipper and Perktold, Josef and others},
+  journal={scipy},
+  volume={7},
+  number={1},
+  pages={92--96},
+  year={2010}
+}
