### 2026-03-13 278db0f
docs: fix statsmodels citation typo and remove duplicate seaborn entry
