### 2026-03-13 042c9f1
docs: formalize 'YData Profiling' capitalization in paper text and table
