### 2026-03-14 25c1495
docs: remove paper source files from main branch (maintained in dev branches)
