### 2026-03-14 0ef993a
docs: remove diagram generation script from main branch
