### 2026-03-14 46fcf25
feat: establish reproducible demo data with generation script
