### 2026-03-14 b08ba48
docs: add OS badge and remove redundant text bullets in README
