### 2026-03-14 9e3c4d3
docs: update project title to include 'Models' for consistency
