### 2026-03-14 ea04596
docs: tighten README header aesthetics and remove redundant spacing
