### 2026-03-14 4f03718
docs: use correct capitalization for YData Profiling in README
