### 2026-03-14 d7636f3
docs: update workflow terminology to be more descriptive
