+++ README.md
@@ -41,7 +41,7 @@ For detailed guides, API reference, and tutorials, visit our official documentat
 
 ## Why TanML?
 
-* **End-to-end workflow:** Data Profiling → Preprocessing → Feature Ranking → Model Development → Evaluation → Reports—all in one UI.
+* **End-to-end workflow:** Data Profiling → Data Preprocessing → Feature Power Ranking → Model Development → Evaluation → Reports—all in one UI.
 * **Audit-ready Word reports:** Generate editable .docx documents for stakeholders and compliance reviews.
 * **Built for regulated industries:** Designed for MRM, credit risk, insurance, and SR 11-7 contexts.
 * **No code required:** Fully UI-driven—no Python knowledge needed.
