### 2026-03-14 7b6897d
docs: include statsmodels in stack list and update Scikit-learn capitalization
