### 2026-03-14 cee540c
docs: capitalize Statsmodels for consistent branding
