+++ docs/index.md
@@ -23,7 +23,7 @@
 * **Built for regulated industries:** Designed for MRM, credit risk, insurance, and SR 11-7 contexts.
 * **No code required:** Fully UI-driven—no Python knowledge needed.
 * **Robust evaluation:** Drift detection, stress testing, SHAP explainability, cluster coverage.
-* **Works with your stack:** Scikit-learn, XGBoost, LightGBM, CatBoost, statsmodels.
+* **Works with your stack:** Scikit-learn, XGBoost, LightGBM, CatBoost, Statsmodels.
 
 ---
 
