### 2026-03-14 47559f2
docs: remove 'Word' from 'Word reports' for more concise terminology
