+++ README.md
@@ -42,7 +42,7 @@ For detailed guides, API reference, and tutorials, visit our official documentat
 ## Why TanML?
 
 * **End-to-end workflow:** Data Profiling → Data Preprocessing → Feature Power Ranking → Model Development → Evaluation → Reports—all in one UI.
-* **Audit-ready Word reports:** Generate editable .docx documents for stakeholders and compliance reviews.
+* **Audit-ready reports:** Generate editable .docx documents for stakeholders and compliance reviews.
 * **Built for regulated industries:** Designed for MRM, credit risk, insurance, and SR 11-7 contexts.
 * **No code required:** Fully UI-driven—no Python knowledge needed.
 * **Robust evaluation:** Drift detection, stress testing, SHAP explainability, cluster coverage.
@@ -211,7 +211,7 @@ $env:TANML_SERVER_ADDRESS="0.0.0.0"; $env:TANML_PORT="9000"; $env:TANML_MAX_MB="
 
 ## Reports
 
-TanML generates audit-ready Word reports (.docx) programmatically:
+TanML generates audit-ready reports (.docx) programmatically:
 
 * **Model Development Report** — Cross-validation metrics, diagnostics, and performance summary
 * **Model Evaluation Report** — Train/Test comparison, drift analysis, stress testing, SHAP explainability
