### 2026-03-14 dfe180c
docs: remove dashes from 'audit ready' for branding consistency
