### 2026-03-14 424b970
docs: explicitly set version to 0.2.0 in mkdocs.yml
