### 2026-03-14 0e5823b
docs: revert hardcoded version, rely on dynamic tag detection for v0.2.0
