### 2026-03-14 1cad024
ci: fix documentation version detection and clear Node.js warnings
