+++ .github/workflows/ci.yml
@@ -16,6 +16,9 @@ on:
       - 'LICENSE'
       - 'CITATION.cff'
 
+env:
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
+
 permissions:
   contents: read
 
