### 2026-03-14 e964b82
docs: trigger build to sync header with official v0.2.0 release
