### 2026-03-14 25c5780
docs: standardize all copyright and citation dates to 2026
