### 2026-03-14 14c36e0
docs: add fast demo GIF and optimized MP4 to README
