### 2026-03-14 5503cf3
docs: replace GIF with high-quality auto-playing MP4 demo
