### 2026-03-14 84a21a7
docs: fix video rendering by using absolute raw URL
