### 2026-03-15 5f19e40
docs: optimize video for GitHub Pages hosting (<10MB)
