### 2026-03-15 fc34fc8
docs: fix video loop by updating tag syntax and cache busting
