### 2026-03-15 539d700
docs: make demo header clickable and add video controls
