### 2026-03-15 cec4af6
docs: add missing cd command to contribution setup instructions
