### 2026-03-15 ad2622f
docs: standardize cd command in Testing section for consistency
