### 2026-03-18 4e45479
docs: split architecture figure into high-level and detailed validation suite views
