### 2026-03-18 386ac34
docs: refine figure description to 'modular overview' in paper.md
