+++ paper.md
@@ -89,7 +89,7 @@ The **Model Development** layer facilitates champion-challenger comparisons usin
 
 The **Evaluation Suite** is the core validation engine. It calculates Population Stability Index (PSI) for drift analysis, tree-based SHAP values [@lundberg2017unified] for explainability, and segmented performance metrics. For example, a model risk team validating an insurance claims model can use PSI to assess whether the production data has drifted from the training distribution, examine SHAP plots to verify that feature contributions align with domain expectations, and review performance breakdowns across customer segments to detect potential fairness or stability issues. These outputs are then serialized directly into the generated validation report.
 
-To illustrate a complete workflow: a quantitative analyst at a financial institution receives a challenger credit model and its associated dataset. Using TanML, the analyst profiles the data to check for quality issues, ranks features by predictive power, retrains the model under cross-validation to verify reported performance, runs drift and SHAP analyses against a holdout or production sample, and exports a structured Word report. This report can then be circulated to model governance committees and regulators with minimal additional preparation. To illustrate the system flow, **Figure \ref{fig:arch}** shows the high-level end-to-end pipeline, while **Figure \ref{fig:validation}** provides a detailed view of the individual analytical modules within the Validation Suite.
+To illustrate a complete workflow: a quantitative analyst at a financial institution receives a challenger credit model and its associated dataset. Using TanML, the analyst profiles the data to check for quality issues, ranks features by predictive power, retrains the model under cross-validation to verify reported performance, runs drift and SHAP analyses against a holdout or production sample, and exports a structured Word report. This report can then be circulated to model governance committees and regulators with minimal additional preparation. To illustrate the system flow, **Figure \ref{fig:arch}** shows the high-level end-to-end pipeline, while **Figure \ref{fig:validation}** provides a modular overview of the specific validation functionalities included in the toolkit.
 
 ![High-level modular workflow of TanML. \label{fig:arch}](architecture_main.png)
 
