### 2026-03-18 9b04610
docs: use visual font size matching (36px macro, 48px micro) in JOSS figures
