### 2026-03-18 44ad509
ci: update JOSS paper trigger paths for new architecture assets
