### 2026-03-19 fe1e334
docs: revert to unified architecture diagram, add Core Architecture intro, update ORCIDs and citations
