+++ paper.bib
@@ -157,3 +157,21 @@
   pages={92--96},
   year={2010}
 }
+
+
+@book{yurdakul2018statistical,
+  title={Statistical properties of population stability index},
+  author={Yurdakul, Bilal},
+  year={2018},
+  publisher={Western Michigan University}
+}
+
+
+@techreport{oecd2020insuranceai,
+  author = {{OECD}},
+  title = {The Impact of Big Data and Artificial Intelligence (AI) in the Insurance Sector},
+  year = {2020},
+  institution = {OECD Publishing},
+  doi = {10.1787/c822ee53-en},
+  url = {https://doi.org/10.1787/c822ee53-en}
+}
