+++ paper.md
@@ -12,10 +12,13 @@ authors:
     orcid: 0009-0004-8583-2208
     affiliation: 1
   - name: Dolly Sah
+    orcid: 0009-0008-9190-9143
     affiliation: 2
   - name: Harshul Jain
+    orcid: 0009-0005-8083-5108
     affiliation: 3
   - name: Tanya Sah
+    orcid: 0000-0002-7366-2318
     affiliation: 3
   - name: Kayden Jordan
     affiliation: 1
@@ -26,7 +29,7 @@ affiliations:
     index: 2
   - name: New York University
     index: 3
-date: 7 March 2026
+date: 18 March 2026
 bibliography: paper.bib
 ---
 
@@ -36,7 +39,9 @@ TanML [@tanml_2026] is an open-source toolkit for validating machine learning mo
 
 # Statement of need
 
-Machine learning models trained on structured, tabular data are widely used in high-stakes domains such as finance, insurance, and other regulated settings. In these environments, building a predictive model is only one part of the workflow, because practitioners must also assess data quality, model behavior, predictive performance, robustness, and interpretability while producing documentation that supports internal review, governance, and compliance processes. For example, banking institutions developing credit scoring or fraud detection models are required to produce comprehensive model validation reports documenting data integrity, performance benchmarking, and sensitivity analyses before regulatory approval [@sr11-7]. Similarly, insurance firms building claims forecasting models must demonstrate that their automated systems are fair, explainable, and robust to distributional shifts. In practice, these validation activities are often fragmented across notebooks, scripts, visualization libraries, and manually prepared reports, making workflows difficult to standardize, reproduce, and communicate, especially when results must be reviewed by stakeholders beyond the original model developer. TanML was developed to address this problem by providing a unified workflow for validating and documenting models trained on tabular data. Instead of requiring practitioners to assemble separate tools for analysis, evaluation, and reporting, TanML organizes these activities within a single interface intended to make validation outputs easier to reproduce, review, and communicate. The primary target audience includes data scientists, quantitative analysts, model validation teams, and other practitioners working in regulated or documentation heavy settings who need reviewable validation artifacts in addition to numerical metrics or exploratory analysis.
+Machine learning models trained on structured, tabular data are widely used in high-stakes domains such as finance, insurance, and other regulated settings. In these environments, building a predictive model is only one part of the workflow, because practitioners must also assess data quality, model behavior, predictive performance, robustness, and interpretability while producing documentation that supports internal review, governance, and compliance processes. For example, banking institutions developing credit scoring or fraud detection models are required to produce comprehensive model validation reports documenting data integrity, performance benchmarking, and sensitivity analyses before regulatory approval [@sr11-7]. Similarly, insurance firms building claims models must demonstrate that their automated systems are fair, explainable, and robust to distributional shifts [@oecd2020insuranceai]. 
+
+In practice, these validation activities are often fragmented across notebooks, scripts, visualization libraries, and manually prepared reports, making workflows difficult to standardize, reproduce, and communicate, especially when results must be reviewed by stakeholders beyond the original model developer. TanML was developed to address this problem by providing a unified workflow for validating and documenting models trained on tabular data. Instead of requiring practitioners to assemble separate tools for analysis, evaluation, and reporting, TanML organizes these activities within a single interface intended to make validation outputs easier to reproduce, review, and communicate. The primary target audience includes data scientists, quantitative analysts, model validation teams, and other practitioners working in regulated or documentation heavy settings who need reviewable validation artifacts in addition to numerical metrics or exploratory analysis.
 
 
 # State of the field
@@ -71,6 +76,8 @@ TanML was explicitly designed as a modular, privacy-first desktop application th
 
 ### Core Architecture
 
[…]
