### 2026-03-19 11c7925
Delete generate_diagram_v2.py
