### 2026-03-19 #280 @tradertanmay
@anabeltan and @echen1214, thank you for reviewing, and @crhea93 thank you for serving as both editor and editor-in-chief.

@anabeltan  As suggested:
1) added a clickable Table of Contents
2) added a new Remote Access (SSH) section. I tested TanML on a Linux server using SSH and VS Code, and auto port forwarding was enabled, which is why I did not catch the issue you pointed out earlier.
3) added cd  tanml in the testing section.

@echen1214
Thank you for your feedback on the UI. I agree with you. To make the UI easier to understand, I made three additions:

a) added a “[See TanML in Action](https://tdlabs-ai.github.io/tanml/assets/tanml_demo.mp4?v=2)” demo by recording the screen and placed the link in the README.
b) added a TanML workflow guide to the Home page.
c) added train and test datasets under the demo_datasets folder.

-The dark mode issue has been fixed. See PR https://github.com/tdlabs-ai/tanml/pull/12
-The workflow links on the main UI are now functional. As shown in the “[See TanML in Action](https://tdlabs-ai.github.io/tanml/assets/tanml_demo.mp4?v=2)” section video, they are working now.
-The Python version is now listed as a badge.
-In the Contributing section, I added the line: git clone https://github.com/tdlabs-ai/tanml.git
-The API reference documentation has also been expanded.

@crhea93  @anabeltan  @echen1214  I am also interested in submitting it to JOSS, so please advise me on what I should do. paper.md, paper.bib, and paper.pdf are in this branch: /dev/finalize-paper.
