### 2026-04-22 #280 @anabeltan
@tradertanmay Thank you for addressing my earlier comments. The updates look good, and I'm satisfied with the changes made. Now that the UI is available via ssh, I was able to explore it more thoroughly, and a few additional questions/issues came up:

1. When uploading your example data, the **Feature Power Ranking** section throws an error. Specifically, the 'Feature Metrics Board', 'Distribution Overlay', and 'Correlation Heatmap' fail with:
```python
AttributeError: 'Styler' object has no attribute 'applymap'
Traceback:
File "/home/usr/anaconda3/envs/test/lib/python3.11/site-packages/tanml/ui/app.py", line 126, in <module>
    main()
File "/home/usr/anaconda3/envs/test/lib/python3.11/site-packages/tanml/ui/app.py", line 118, in main
    render_feature_ranking_page(run_dir)
File "/home/usr/anaconda3/envs/test/lib/python3.11/site-packages/tanml/ui/views/ranking.py", line 422, in render_feature_ranking_page
    .applymap(color_p_value, subset=["p-value"])
     ^^^^^^^^
```
This is not a bug in logic, but it's a compatibility issue with newer pandas. This line uses `Styler.applymap`, which has been removed in recent versions of pandas (≥ 2.1). This will raise an `AttributeError` in newer environments.

This could be fixed by replacing:

```python
.applymap(color_p_value, subset=["p-value"])
```

with:

```python
.map(color_p_value, subset=["p-value"])
```

Alternatively, if this tool is intended to support multiple pandas versions, you may want to add a compatibility check (e.g., fallback to `applymap` when `map` is unavailable).

It would also help to pin or document the expected pandas version in the project dependencies to avoid this kind of breakage.

2. The tool itself is strong, but the documentation [here](https://tdlabs-ai.github.io/tanml/user-guide/getting-started/) feels a bit too high-level. It would be helpful to expand on what each section (or in your words, 'step') does, what functionality is available within them, which options (e.g., libraries in Model Config under Model Development) are best suited for what use cases, what outputs users should expect and how they can help with the user's work, and so on.

Overall, this is shaping up nicely :) Looking forward to seeing it released.
