### 2026-04-30 c2b13dc
Update getting-started.md
