+++ docs/user-guide/getting-started.md
@@ -25,19 +25,217 @@ tanml ui
 ```
 
 ### Step 1: Understand Your Data (Data Profiling)
-Start by uploading your raw dataset (**CSV** or **Parquet** format). Once you select your **Target Variable**, TanML will automatically analyze your data quality onscreen, identifying missing values, highlighting outliers, and visualizing the distribution of each variable to provide a baseline understanding of your data.
+
+#### Overview  
+Start by uploading your raw dataset. TanML supports a wide range of file formats, including:
+
+- **Standard formats:** CSV, TSV, Excel (.xlsx, .xls), JSON, TXT  
+- **High-performance formats:** Parquet  
+- **Statistical/legacy formats:** Stata (.dta), SPSS (.sav), SAS (.sas7bdat)  
+
+TanML performs automated exploratory data analysis directly within the interface to assess data quality and statistical structure. This step is **target-agnostic**, meaning it helps you identify issues across the entire dataset before you even decide what you intend to predict.
+
+---
+
+#### Functionalities  
+
+**Missing Value Analysis**  
+TanML computes the percentage of missing values for each feature and provides visual bar charts to highlight gaps. This helps identify features that require imputation or removal in Step 2.
+
+**Distribution Analysis**  
+The tool visualizes the distribution of numerical variables using **Histograms** and **Box Plots**. This allows users to detect skewness, median shifts, and the spread of data.
+
+**Outlier Identification**  
+TanML uses statistical IQR thresholds to surface extreme observations. Users can also set **custom domain constraints** (Min/Max) to identify data points that violate business logic.
+
+**Duplicate Detection**  
+Users can identify identical records based on a custom subset of columns. This is critical for ensuring data integrity and preventing "leakage" before the data is split.
+
+**Automated Risk Flags**  
+TanML automatically flags columns with high cardinality (too many unique categories) or near-constant values, warning you of features that might cause problems for machine learning models later.
+
+---
+
+#### Outputs  
+- **Interactive Visualizations**: Histograms, Box Plots, and Correlation Heatmaps.  
+- **Summary Statistics**: Automated calculation of Mean, Median, Std Dev, and unique counts.  
+- **Downloadable Reports**: Export specific lists of identified **Outliers** or **Duplicates** as CSV files for manual review.
+
+---
+
+#### Why This Step Matters  
+This step establishes a "Health Check" for your data. It enables you to catch "Garbage In, Garbage Out" scenarios early, ensuring that your preprocessing decisions in the next step are based on a solid statistical understanding of your raw data. 
 
 ### Step 2: Clean & Split (Data Preprocessing)
-Raw data often requires preparation before modeling. Upload your dataset to apply imputation strategies for missing values and encode text categories. Then, use the built-in **Data Splitter** at the bottom of the page to divide your cleaned dataset into a dedicated *Training Set* (for model learning) and an independent *Testing Set* (for final validation). You can export and download these datasets for the next steps.
+
+#### Overview  
+This step prepares the dataset for modeling by handling missing values, encoding categorical variables, and creating separate training and testing datasets. TanML ensures that every transformation is logged and reproducible.
+
+---
+
+#### Functionalities  
+
+**Advanced Imputation**  
+TanML provides several strategies for handling gaps in your data:
+- **Numeric**: Choose from `Mean`, `Median`, `Zero`, or the advanced **KNN (K-Nearest Neighbors)** imputer.
+- **Categorical**: Apply `Mode` (Most Frequent) or create an explicit **"Missing_Label"** category.
+- **Row Deletion**: Use the **"Drop Rows"** option to remove records with missing data entirely.
+
[…]
