### 2026-04-30 #280 @tradertanmay
@anabeltan Thank you for the review. Please let me know if this is ready for approval from your side.

The requested changes have been implemented:

1) Replaced applymap with map (PR: https://github.com/tdlabs-ai/tanml/pull/14).
2) Updated the documentation [here](https://tdlabs-ai.github.io/tanml/user-guide/getting-started/)
