### 2026-04-30 #280 @echen1214
@tradertanmay thanks for the progress, I agree with @anabeltan it seems like the tool is shaping up well. Thanks for addressing my concerns, the UI is improved and the video is appreciated. During my usage, however
1. I do not see the side bar anymore. I am running the latest version, tanml-0.2.0 locally

2. I also like the remote SSH suggestion and can confirm it works in my VS code IDE, although still without the side bar
