### 2026-04-30 #280 @tradertanmay
@echen1214 Thanks for the review. I’ve updated the code to address the issue.

You can test it locally with:
git clone https://github.com/tdlabs-ai/tanml.git
cd tanml
pip install -e .
tanml ui

Issue:
This issue generally occurred when the user minimized the sidebar.

Fix:
Removed `visibility: hidden` from `tanml/ui/config.py`.

Please let me know if this is ready for approval from your side.
