### 2026-09-03 #280 @crhea93
@tradertanmay Congrats! Can you please add a link to the JOSS submission issue here when you submit. Can you also please provide the final Zenodo DOI once you have it?
