### 2026-09-04 #280 @tradertanmay
@crhea93  @anabeltan  @echen1214 Thank you for reviewing this and for your time. Without your support, I would not have been able to catch those bugs.

@crhea93
- **JOSS Submission Issue**: https://github.com/openjournals/joss-reviews/issues/11251
- **Final Zenodo DOI**: https://zenodo.org/records/22285793
- completed the maintainer post-review survey.
