### 2026-09-04 #280 @tradertanmay
@crhea93 @lwasser  @echen1214 @anabeltan Once again, thank you for reviewing, and the paper just got accepted into JOSS.
